feat:es_demo 完成

+++ src/main/java/com/gyc/es/bean/BookSearchRequest.java
@@ -7,6 +7,6 @@ public class BookSearchRequest extends SearchRequest {
 
     private String keyword;
 
-    private Double[] vector;
+    private float[] vector;
 
 }
